+++ src/huffman.h
@@ -1,6 +1,7 @@
 #ifndef HUFFMAN_H
 #define HUFFMAN_H
+#include <stdlib.h>
 
-void huffman( char *stringToCompress, unsigned int stringLength );
+void huffman( char *stringToCompress, size_t stringLength );
 
 #endif
